bcachefs: Improve bch2_insert_snapshot_whiteouts()
This improves the check for whether @new_pos needs a whiteout:
pos_overwitten_in_snapshot() will return true if there's already a
whiteout present.

This lets us avoid emitting whiteouts that already exist. Previously we
were using bch2_btree_iter_peek_slot() in BTREE_ITER_FILTER_SNAPSHOTS
mode, which would return a hole if it found a whiteout.

Since bch2_insert_snapshot_whiteouts() may have to insert an unbounded
number of snapshots we're going to need to make it run incrementally in
the future - this is a step in that direction.

Signed-off-by: Kent Overstreet <kent.overstreet@linux.dev>